requirements.txt
cycler==0.11.0
hdbscan==0.8.30
iminuit==2.6.1
joblib==1.3.0
matplotlib==3.6.2
numba==0.57.1
numpy==1.23.5
pandarallel==1.6.5
pandas==1.5.2
scikit_learn==1.0.2
scipy==1.9.3
seaborn==0.12.2
tqdm==4.65.0
umap_learn==0.5.3
xarray==2023.6.0
pyarrow==11.0.0